+++ tests/test_logical.py
@@ -1,12 +1,7 @@
 from pyscipopt import Model
 from pyscipopt import quicksum
 
-try:
-    import pytest
-    itertools = pytest.importorskip("itertools")
-except ImportError:
-    import itertools
-product = itertools.product
+import pytest
 
 
 # Testing AND/OR/XOR constraints