+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 - Fixed the type of @ matrix operation result from MatrixVariable to MatrixExpr.
 ### Changed
+- Add a PEP 735 dependency group for test dependencies in `pyproject.toml`
 ### Removed
 
 ## v5.6.0 - 2025.08.26